Additional file splitting and dependency addition to fix netstandard/netfx build problems
