+++ src/Common/src/CoreLib/System/IO/Path.cs
@@ -905,33 +905,5 @@ namespace System.IO
             IsCaseSensitive ?
                 StringComparison.Ordinal :
                 StringComparison.OrdinalIgnoreCase;
-
-        /// <summary>
-        /// Trims one trailing directory separator beyond the root of the path.
-        /// </summary>
-        public static string TrimEndingDirectorySeparator(string path) =>
-            EndsInDirectorySeparator(path) && !PathInternal.IsRoot(path.AsSpan()) ?
-                path.Substring(0, path.Length - 1) :
-                path;
-
-        /// <summary>
-        /// Trims one trailing directory separator beyond the root of the path.
-        /// </summary>
-        public static ReadOnlySpan<char> TrimEndingDirectorySeparator(ReadOnlySpan<char> path) =>
-            EndsInDirectorySeparator(path) && !PathInternal.IsRoot(path) ?
-                path.Slice(0, path.Length - 1) :
-                path;
-
-        /// <summary>
-        /// Returns true if the path ends in a directory separator.
-        /// </summary>
-        public static bool EndsInDirectorySeparator(ReadOnlySpan<char> path)
-            => path.Length > 0 && PathInternal.IsDirectorySeparator(path[path.Length - 1]);
-
-        /// <summary>
-        /// Returns true if the path ends in a directory separator.
-        /// </summary>
-        public static bool EndsInDirectorySeparator(string path)
-              => !string.IsNullOrEmpty(path) && PathInternal.IsDirectorySeparator(path[path.Length - 1]);
     }
 }
